+++ src/KMonad/Keyboard/ComposeSeq.hs
@@ -20,6 +20,7 @@ module KMonad.Keyboard.ComposeSeq
 where
 
 import KMonad.Prelude
+import RIO.Text.Partial (replace)
 
 --------------------------------------------------------------------------------
 
@@ -30,13 +31,20 @@ import KMonad.Prelude
 -- 3. A descriptive-name
 --
 ssComposed :: [(Text, Char, Text)]
-ssComposed =
+ssComposed = composeSeqs & (each . _1) %~ sanitize
+ where
+  sanitize :: Text -> Text
+  sanitize = replace "(" "\\("
+           . replace ")" "\\)"
+           . replace "_" "\\_"
+
+  composeSeqs :: [(Text, Char, Text)]
+  composeSeqs =
     [ ("' '"      , '´'     , "acute")
     , ( "^ -"     , '¯'     , "macron" )
     , ( "spc ("   , '˘'     , "breve" )
     , ( "\" \""   , '¨'     , "diaeresis" )
     , ("spc <"    , 'ˇ'     , "caron")
-    , ("` spc"    , '`'     , "grave")
     , (", spc"    , '¸'     , "cedilla")
     , ("spc spc"  , ' '     , "nobreakspace")
     , ("spc ."    , ' '     , "U2008")
@@ -733,6 +741,7 @@ ssComposed =
     , ("> _"      , '≥'     , "U2265")
 
     -- Sequences that should exist but do not work
+    --, ("` spc"    , '`'     , "grave") -- recursive and incorrect. It's <dead_grave> <space> and <dead_grave> is not mapped in en_US
     --, ("^ spc", '^', "asciicircum") -- This overlaps with the normal 'shifted-6' macro for
     -- , ("' j", 'j́', "jacute")
     ]