+++ changelog.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 
 ### Fixed
 
+- Fixed broken compose sequences (#823)
+
 ## 0.4.2 – 2023-10-07
 
 ### Added