+++ changelog.md
@@ -8,11 +8,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 
 ### Added
 
+- Added `stepped`. It performs the next button from a circular sequence
+  whenever it is pressed.
+- Implemented named source blocks.
+  To name a source block add `:name <name>` at the beginning of the
+  `defsrc` block. To use it add `:source <name>` after the layer name to the
+  `deflayer` block. (#831)
+
 ### Changed
+
+- Update Karabiner-DriverKit to 3.1.0 (#780)
+- Added tests to check that every button has documentation (#857)
 
 ### Fixed
 
+- Fixed crash on non-US backslash under MacOS (#766)
+- Fixed broken keyboard due to circular event handeling under MacOS (#781)
+- Fixed crash on unhandled buttons by ignoring them (#807)
+- Fixed parse errors relating to whitespace (#796)
 - Fixed broken compose sequences (#823)
+- Fixed parse errors when using keys only available on darwin os (#828)
+- Fixed `around-next` wasn't parsable (#857)
 
 ## 0.4.2 – 2023-10-07