+++ changelog.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
   To name a source block add `:name <name>` at the beginning of the
   `defsrc` block. To use it add `:source <name>` after the layer name to the
   `deflayer` block. (#831)
+- Added `around-only`, which works like `around` but releases the outer button
+  when others are pressed.
+- Added `around-when-alone`, which also represses the outer button when only
+  the inner button is pressed and all others have been released.
 
 ### Changed